% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{anySpecial}
\alias{anySpecial}
\title{Detect whether there are any 'specials' in a formula term}
\usage{
anySpecial(term, specials = findReTrmClasses())
}
\arguments{
\item{term}{formula term}

\item{specials}{values to detect}
}
\value{
logical value
}
\description{
Detect whether there are any 'specials' in a formula term
}
