% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chck1}
\alias{chck1}
\title{Check for constant terms.}
\usage{
chck1(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
NULL.  The function is executed for its side effect.
}
\description{
Check for a constant term (a literal 1) in an expression
In the mixed-effects part of a nonlinear model formula, a constant
term is not meaningful because every term must be relative to a
nonlinear model parameter.  This function recursively checks the
expressions in the formula for a a constant, calling stop() if
such a term is encountered.
}
