% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expandGrpVar}
\alias{expandGrpVar}
\title{apply operator expansion (e.g. a/b -> a + a:b) to a formula term}
\usage{
expandGrpVar(f)
}
\arguments{
\item{f}{a language object (an atom of a formula)}
}
\description{
apply operator expansion (e.g. a/b -> a + a:b) to a formula term
}
\examples{
expandGrpVar(quote(x*y))
expandGrpVar(quote(x/y))
}
