% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarCorr.R
\name{format_sdvar}
\alias{format_sdvar}
\alias{format_corr}
\alias{get_sd}
\title{Format columns corresponding to std. dev. and/or variance.}
\usage{
format_sdvar(
  reStdDev,
  use.c = "Std.Dev.",
  formatter = format,
  digits = max(3, getOption("digits") - 2),
  ...
)

format_corr(x, maxdim = Inf, digits = 2, maxlen = 10, ...)

get_sd(x, ...)
}
\arguments{
\item{reStdDev}{a vector of standard deviations.}

\item{use.c}{a character vector indicating which scales to include.}

\item{formatter}{formatting function.}

\item{digits}{digits for format}

\item{...}{additional parameters}

\item{x}{a square numeric matrix}

\item{maxdim}{maximum number of rows/columns to display}

\item{maxlen}{maximum number of rows to display}
}
\description{
Format columns corresponding to std. dev. and/or variance.
}
