% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_grpvars}
\alias{get_grpvars}
\title{get grouping variable symbols/names}
\usage{
get_grpvars(formula, return_val = c("char", "symbol"))
}
\arguments{
\item{formula}{a formula}

\item{return_val}{return character string or raw symbol?}
}
\description{
get grouping variable symbols/names
}
\examples{
form <- Reaction ~ Days + (Days | group / Subject)
get_grpvars(form)
get_grpvars(form, return_val = "symbol")
}
