% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randint.R
\name{randint}
\alias{randint}
\title{Remove all random slopes from a formula, while retaining random intercepts.}
\usage{
randint(form)
}
\arguments{
\item{form}{A formula}
}
\value{
The new formula
}
\description{
Remove all random slopes from a formula, while retaining random intercepts.
}
\examples{
f <- ~ 1 + a  + b + (a | f) + (1 + a | g) + (a + b | h ) + (1 + a + b | i)
randint(f)
}
