% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{RHSForm}
\alias{RHSForm}
\title{extract right-hand side of a formula}
\usage{
RHSForm(form, as.form = FALSE)
}
\arguments{
\item{form}{a formula object}

\item{as.form}{(logical) return a formula (TRUE) or as a call/symbolic object (FALSE) ?}
}
\value{
a \code{language} object
}
\description{
extract right-hand side of a formula
}
\examples{
RHSForm(y ~ x + (1|g))
}
